+++ src/__mocks__/expo-modules-core.js
@@ -19,6 +19,7 @@ module.exports = {
     getPendingTransactionsIOS: jest.fn(),
     clearTransactionIOS: jest.fn(),
     // Common methods
+    fetchProducts: jest.fn(),
     requestProducts: jest.fn(),
     requestPurchase: jest.fn(),
     requestPurchaseOnPromotedProductIOS: jest.fn(),